refactor(policy): extract lease init into a separate module
In preparation for updating kubert to use a newer Lease API (supporting
diagnostics), this change extracts the lease initialization logic into a
dedicated module so that the new interface may be established (roughly) before
the details change.
